move events query
